requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.2.1
django-background-tasks==1.2.8
django-celery-beat==2.8.1
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
dotenv==0.9.9
drf-yasg==1.21.10
h11==0.16.0
idna==3.10
inflection==0.5.1
kombu==5.5.3
numpy==2.2.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
prompt_toolkit==3.0.51
psycopg==3.2.9
psycopg2-binary==2.9.10
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
requests==2.32.3
selenium==4.32.0
six==1.17.0
slugify==0.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
